# backend/api-docs.yaml
openapi: 3.1.1
info:
  title: Rat Hacks API
  version: 1.0.0
  description: API documentation for Rat Hacks
servers:
  - url: https://rathacks.com
    description: Rat Hacks Server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*]).{8,16}$'
                firstName:
                  type: string
                lastName:
                  type: string
                homeSchool:
                  type: string
                gradeLevel:
                  type: string
                  enum: ["nine", "ten", "eleven", "twelve"]
                isGovSchool:
                  type: boolean
                techStacK:
                  type: string
                previousHackathon:
                  type: boolean
                parentFirstName:
                  type: string
                parentLastName:
                  type: string
                parentEmail:
                  type: string
                  format: email
                parentPhoneNumber:
                  type: string
                  example: 5408531056
                contactFirstName:
                  type: string
                contactLastName:
                  type: string
                contactRelationship:
                  type: string
                contactPhoneNumber:
                  type: string
                  example: 5408531056
              required:
                - email
                - password
                - firstName
                - lastName
                - homeSchool
                - gradeLevel
                - isGovSchool
                - techStacK
                - previousHackathon
                - parentFirstName
                - parentLastName
                - parentEmail
                - parentPhoneNumber
                - contactFirstName
                - contactLastName
                - contactRelationship
                - contactPhoneNumber
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                required:
                  - message
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All fields are required
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to register user
                required:
                  - message
  /api/auth/verify-email:
    post:
      summary: Verify a user's email
      tags:
        - Auth
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          description: User's Email
        - name: token
          in: token
          required: true
          schema:
            type: string
          description: User's token to verify email
      responses:
        200:
          description: Email successfully verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email verified successfully
                required:
                  - message
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid email or token
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to verify email
                required:
                  - message
  /api/auth/resend-verification:
    post:
      summary: Resend the verification email
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: Resent the verification email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification email resent
                required:
                  - message
        400:
          description: Email is already verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email is already verified
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        429:
          description: Too many requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: You can only request a new verification email every 15 minutes
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to resend verification email
                required:
                  - message
  /api/auth/login:
    post:
      summary: Login a user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  token:
                    type: string
                    example: yourToken123
                    format: jwt
                required:
                  - message
                  - token
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email and password are required
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid email or password
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to login user
                required:
                  - message
  /api/auth/logout:
    post:
      summary: Logout the current user
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properites:
                  message:
                    type: string
                    example: Failed to logout user
                required:
                  - message
  /api/auth/logout-all:
    post:
      summary: Logout the current user from all devices
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: User logged out from all devices successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout from all devices
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to logout user from all devices
                required:
                  - message
  /api/auth/current-user:
    get:
      summary: Get the currently logged in user
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: successfully got the current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                    format: email
                  emailVerified:
                    type: boolean
                  accountType:
                    type: string
                    enum: ["student", "organizer", "judge"]
                  firstName:
                    type: string
                  lastName:
                    type: string
                  schoolDivision:
                    type: string
                  gradeLevel:
                    type: string
                    enum: ["nine", "ten", "eleven", "twelve"]
                  isGovSchool:
                    type: boolean
                  techStack:
                    type: string
                  previousHackathon:
                    type: boolean
                  parentFirstName:
                    type: string
                  parentLastName:
                    type: string
                  parentEmail:
                    type: string
                    format: email
                  parentPhoneNumber:
                    type: string
                    example: 5408531056
                  contactFirstName:
                    type: string
                  contactLastName:
                    type: string
                  contactRelationship:
                    type: string
                  contactPhoneNumber:
                    type: string
                    example: 5408531056
                  events:
                    type: array
                    items:
                      type: string
                required:
                  - id
                  - email
                  - isEmailVerified
                  - accountType
                  - firstName
                  - lastName
                  - schoolDivision
                  - gradeLevel
                  - isGovSchool
                  - techStacK
                  - previousHackathon
                  - parentFirstName
                  - parentLastName
                  - parentEmail
                  - parentPhoneNumber
                  - contactFirstName
                  - contactLastName
                  - contactRelationship
                  - contactPhoneNumber
                  - events
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal Server Error
                required:
                  - message
  /api/auth/update:
    put:
      summary: Update the current user's information
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                schoolDivision:
                  type: string
                gradeLevel:
                  type: string
                  enum: ["nine", "ten", "eleven", "twelve"]
                isGovSchool:
                  type: boolean
                techStack:
                  type: string
                previousHackathon:
                  type: boolean
                parentFirstName:
                  type: string
                parentLastName:
                  type: string
                parentEmail:
                  type: string
                  format: email
                parentPhoneNumber:
                  type: string
                  example: 5408531056
                contactFirstName:
                  type: string
                contactLastName:
                  type: string
                contactRelationship:
                  type: string
                contactPhoneNumber:
                  type: string
              required:
                - firstName
                - lastName
                - schoolDivision
                - gradeLevel
                - isGovSchool
                - techStack
                - previousHackathon
                - parentFirstName
                - parentLastName
                - parentEmail
                - parentPhoneNumber
                - contactFirstName
                - contactLastName
                - contactRelationship
                - contactPhoneNumber
      responses:
        200:
          description: User information updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User information updated successfully
                required:
                  - message
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All fields are required
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to update user
                required:
                  - message
  /api/auth/update-password:
    put:
      summary: Update the current user's password
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  minLength: 8
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*]).{8,16}$'
              required:
                - newPassword
      responses:
        200:
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: password updated successfully
                required:
                  - message
          400:
            description: Bad request
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: New password is required
                  required:
                    - message
          401:
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Unauthorized'
                  required:
                    - message
          404:
            description: User not found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: User not found
                  required:
                    - message
          500:
            description: Internal Server Error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Failed to update password
                  required:
                    - message
  /api/auth/delete:
    delete:
      summary: Delete the current user's account
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: User account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User account deleted successfully
                required:
                  - message
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  mesasge:
                    type: string
                    example: Unauthorized
                required:
                  - message
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                required:
                  - message
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to delete user
                required:
                  - message
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token